+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Updated
+
+- Using OnPush as changeDetection mechanism into ngx-skeleton-loader component
+
 ## [2.6.2][] - 2020-12-08
 
 ### Fixed
@@ -335,7 +339,5 @@ Now we can define the animation we want to use in `<ngx-skeleton-loader>` compon
 [2.6.0]: https://github.com/willmendesneto/ngx-skeleton-loader/tree/v2.6.0
 [unreleased]: https://github.com/willmendesneto/ngx-skeleton-loader/compare/v2.6.1...HEAD
 [2.6.1]: https://github.com/willmendesneto/ngx-skeleton-loader/tree/v2.6.1
-
-
-[Unreleased]: https://github.com/willmendesneto/ngx-skeleton-loader/compare/v2.6.2...HEAD
+[unreleased]: https://github.com/willmendesneto/ngx-skeleton-loader/compare/v2.6.2...HEAD
 [2.6.2]: https://github.com/willmendesneto/ngx-skeleton-loader/tree/v2.6.2